mCollect Auto suggestion
In mCollect form, tab on Service Type dropdown there is selection through arrow keys and scrollable but not Typeable (Autocomplete) as available in Service Category dropdown. Tried to type into the service type dropdown for selection of services. Typeable (Autocomplete)  dropdown list is effective in speed along with the arrow keys or scroll button